applied changes - refactoring
