+++ src/pages/index.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import { PageProps, graphql } from "gatsby";
-import Navbar from "../components/Navbar";
 import Hero from "../components/Hero";
 import Footer from "../components/Footer";
 import Features from "../components/Features";
@@ -8,6 +7,7 @@ import Testimonials from "../components/Testimonial";
 import FeaturedBlogs from "../components/Blog";
 import Contact from "../components/Contact";
 import Projects from "../components/Projects";
+import Layout from "../components/layout";
 
 export type IndexPageType = Pick<Queries.IndexPageQuery, "indexPage">;
 
@@ -35,17 +35,14 @@ export const IndexPageTemplate = ({ indexPage, blogs, projects }: {indexPage: In
   blogs.edges.map((item) => (console.log(item.blog.frontmatter?.title)))
   return (
     <>
-      <Navbar />
-
-      <main>
-        <Hero {...hero!} />
-        <Features {...services!} />
-        <Projects {...projects} />
-        <Testimonials {...testimonials!} />
-        {/* <FeaturedBlogs {...blogs}/> */}
-        <Contact {...contact!} />
-      </main>
-
+      <Layout>
+          <Hero {...hero!} />
+          <Features {...services!} />
+          <Projects {...projects} />
+          <Testimonials {...testimonials!} />
+          {/* <FeaturedBlogs {...blogs}/> */}
+          <Contact {...contact!} />
+      </Layout>
       <Footer contact={contact!} />
     </>
   );
