+++ static/admin/config.yml
@@ -62,7 +62,7 @@ collections:
       - { label: "Project Title", name: "title", widget: "string" }
       - { label: "Project Slug", name: "slug", widget: "string" }
       - { label: "Image", name: "image", widget: "image"}
-      - { label: "Description", name: "description", widget: "string" }
+      - { label: "Description", name: "description", widget: "text" }
       - { label: "Body", name: "body", widget: "markdown" }
 
   - name: "pages"
